requirements.txt
absl-py==2.1.0
accelerate==0.28.0
aiohttp==3.9.3
aiosignal==1.3.1
albumentations==0.4.3
altair==5.2.0
antlr4-python3-runtime==4.8
asttokens @ file:///opt/conda/conda-bld/asttokens_1646925590279/work
async-timeout==4.0.3
attrs==23.2.0
backcall @ file:///home/ktietz/src/ci/backcall_1611930011877/work
backports.zoneinfo==0.2.1
blinker==1.7.0
cachetools==5.3.2
certifi==2024.2.2
charset-normalizer==3.3.2
click==8.1.7
-e git+https://github.com/openai/CLIP.git@dcba3cb2e2827b402d2701e7e1c7d9fed8a20ef1#egg=clip
comm @ file:///croot/comm_1709322850197/work
contourpy==1.1.1
cycler==0.12.1
datasets==2.18.0
debugpy @ file:///croot/debugpy_1690905042057/work
decorator @ file:///opt/conda/conda-bld/decorator_1643638310831/work
diffusers==0.16.0
dill==0.3.8
einops==0.3.0
executing @ file:///opt/conda/conda-bld/executing_1646925071911/work
filelock==3.13.1
flask==3.0.3
fonttools==4.55.8
frozenlist==1.4.1
fsspec==2024.2.0
ftfy==6.1.3
future==0.18.3
gitdb==4.0.11
GitPython==3.1.42
google-auth==2.28.0
google-auth-oauthlib==1.0.0
grpcio==1.60.1
huggingface-hub==0.22.0
idna==3.6
imageio==2.9.0
imageio-ffmpeg==0.4.2
imgaug==0.2.6
importlib-metadata @ file:///croot/importlib_metadata-suite_1704813515092/work
importlib-resources==6.1.1
ipykernel @ file:///croot/ipykernel_1705933831282/work
ipython @ file:///croot/ipython_1691532092695/work
itsdangerous==2.2.0
jedi @ file:///tmp/build/80754af9/jedi_1644315233700/work
Jinja2==3.1.3
joblib==1.3.2
jsonschema==4.21.1
jsonschema-specifications==2023.12.1
jupyter-client @ file:///croot/jupyter_client_1699455897726/work
jupyter-core @ file:///croot/jupyter_core_1718818295206/work
kiwisolver==1.4.7
kornia==0.7.2
kornia-rs==0.1.3
-e git+git@github.com:chronos123/VQ_generation.git@d6c862c3ddafd6201dc6ce17ae4de081431cc670#egg=latent_diffusion
lazy-loader==0.3
lpips==0.1.4
Markdown==3.5.2
markdown-it-py==3.0.0
MarkupSafe==2.1.5
matplotlib==3.7.5
matplotlib-inline @ file:///opt/conda/conda-bld/matplotlib-inline_1662014470464/work
mdurl==0.1.2
mkl-fft==1.3.1
mkl-random @ file:///tmp/build/80754af9/mkl_random_1626186064646/work
mkl-service==2.4.0
mpmath==1.3.0
multidict==6.0.5
multiprocess==0.70.16
nest-asyncio @ file:///croot/nest-asyncio_1708532673751/work
networkx==3.1
numpy==1.24.4
nvidia-cublas-cu12==12.1.3.1
nvidia-cuda-cupti-cu12==12.1.105
nvidia-cuda-nvrtc-cu12==12.1.105
nvidia-cuda-runtime-cu12==12.1.105
nvidia-cudnn-cu12==8.9.2.26
nvidia-cufft-cu12==11.0.2.54
nvidia-curand-cu12==10.3.2.106
nvidia-cusolver-cu12==11.4.5.107
nvidia-cusparse-cu12==12.1.0.106
nvidia-nccl-cu12==2.19.3
nvidia-nvjitlink-cu12==12.3.101
nvidia-nvtx-cu12==12.1.105
oauthlib==3.2.2
omegaconf==2.1.1
opencv-python==4.1.2.30
opencv-python-headless==4.9.0.80
packaging==21.3
pandas==2.0.3
parso @ file:///opt/conda/conda-bld/parso_1641458642106/work
peft==0.10.0
pexpect @ file:///tmp/build/80754af9/pexpect_1605563209008/work
pickleshare @ file:///tmp/build/80754af9/pickleshare_1606932040724/work
pillow @ file:///croot/pillow_1707233021655/work
pkgutil-resolve-name==1.3.10
platformdirs @ file:///croot/platformdirs_1692205439124/work
prompt-toolkit @ file:///croot/prompt-toolkit_1704404351921/work
protobuf==3.20.3
psutil==5.9.8
ptyprocess @ file:///tmp/build/80754af9/ptyprocess_1609355006118/work/dist/ptyprocess-0.7.0-py2.py3-none-any.whl
pudb==2019.2
pure-eval @ file:///opt/conda/conda-bld/pure_eval_1646925070566/work
pyarrow==15.0.0
pyarrow-hotfix==0.6
pyasn1==0.5.1
pyasn1-modules==0.3.0
pydeck==0.8.1b0
pyDeprecate==0.3.1
pygments==2.17.2
pyparsing==3.1.2
python-dateutil==2.8.2
pytorch-lightning==1.6.2
pytorch-wavelets==1.3.0
pytz==2024.1
PyWavelets==1.4.1
PyYAML==6.0.1
pyzmq @ file:///croot/pyzmq_1705605076900/work
referencing==0.33.0
regex==2023.12.25
requests==2.31.0
requests-oauthlib==1.3.1
rich==13.7.0
rpds-py==0.18.0
rsa==4.9
sacremoses==0.1.1
safetensors==0.4.2
scikit-image==0.20.0
scipy==1.9.1
six @ file:///tmp/build/80754af9/six_1644875935023/work
smmap==5.0.1
-e git+git@github.com:chronos123/VQ_generation.git@d6c862c3ddafd6201dc6ce17ae4de081431cc670#egg=spherenet&subdirectory=ldm/modules/SphereNet
stack-data @ file:///opt/conda/conda-bld/stack_data_1646927590127/work
streamlit==1.31.1
sympy==1.12
-e git+git@github.com:chronos123/VQ_generation.git@d6c862c3ddafd6201dc6ce17ae4de081431cc670#egg=taming_transformers&subdirectory=src/taming-transformers
tenacity==8.2.3
tensorboard==2.14.0
tensorboard-data-server==0.7.2
test-tube==0.7.5
tifffile==2023.7.10
timm==1.0.8
tokenizers==0.13.3
toml==0.10.2
toolz==0.12.1
torch==2.2.1
torch-fidelity==0.3.0
torchmetrics==0.6.2
torchvision==0.17.1
tornado==6.4
tqdm==4.66.2
traitlets @ file:///croot/traitlets_1718227057033/work
transformers==4.28.1
triton==2.2.0
typing-extensions @ file:///croot/typing_extensions_1705599297034/work
tzdata==2024.1
tzlocal==5.2
urllib3==2.2.1
urwid==2.6.1
validators==0.22.0
watchdog==4.0.0
wcwidth==0.2.13
werkzeug==3.0.1
xxhash==3.4.1
yarl==1.9.4
zipp @ file:///croot/zipp_1704206909481/work
